move container tests in containers folder
